# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Nightshift API
  version: "2.1.0"
  summary: Queue messages for Claude, OpenAI, Gemini and Ollama, to send later or when a usage limit resets.
  description: |
    Everything the Nightshift web app does goes through this API. The browser uses a
    session cookie; other tools use an API token created in Settings.

    Request bodies accept snake_case or camelCase. Responses are camelCase.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1:8787
    description: Default local install
security:
  - bearerToken: []
...
paths:
...
  /v1/models:
    get:
      tags: [Service]
      summary: Models the provider reports
      parameters:
        - { name: provider, in: query, required: true, schema: { $ref: "#/components/schemas/ProviderId" } }
        - { name: refresh, in: query, schema: { type: boolean }, description: "Fetch again instead of using the ten-minute cache" }
      responses:
        "502": { description: "The provider could not be reached and nothing was cached", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "200":
          description: Model list, empty when the provider has no key
          headers:
            X-Models-Cached-At: { schema: { type: integer }, description: "When the list was fetched, epoch ms" }
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string }
                    name: { type: string }
  /v1/limits/{provider}/clear:
    post:
      tags: [Service]
      summary: Declare a limit over and release everything waiting on it
      parameters:
        - { name: provider, in: path, required: true, schema: { $ref: "#/components/schemas/ProviderId" } }
      responses:
        "200": { description: "Cleared" }
  /v1/conversations:
    get:
      tags: [Conversations]
      summary: List conversations
      parameters:
        - { name: provider, in: query, schema: { $ref: "#/components/schemas/ProviderId" } }
      responses:
        "200":
          description: Newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Conversation" }
    post:
      tags: [Conversations]
      summary: Create a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider]
              properties:
                provider: { $ref: "#/components/schemas/ChatProviderId" }
                title: { type: string }
                model: { type: string }
                system: { type: string, description: "System prompt used for the whole conversation" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
  /v1/conversations/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [Conversations]
      summary: One conversation with its messages
      responses:
        "200":
          description: Conversation and history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Conversation"
                  - type: object
                    properties:
                      messages:
                        type: array
                        items: { $ref: "#/components/schemas/StoredMessage" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Conversations]
      summary: Rename, or change the model or system prompt
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                model: { type: string }
                system: { type: string }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
    delete:
      tags: [Conversations]
      summary: Delete it and cancel anything queued for it
      responses:
        "200": { description: "Deleted" }
...
  /v1/sessions/{id}:
    get:
      tags: [Sessions]
      summary: Transcript of one Claude Code session
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: "Session transcript" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/sessions/{id}/export:
    get:
      tags: [Import and export]
      summary: Download a Claude Code session
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: format, in: query, schema: { type: string, enum: [md, json, jsonl], default: md } }
        - { name: tz, in: query, schema: { type: string }, description: "IANA timezone for Markdown timestamps" }
      responses:
        "200":
          description: The transcript
          content:
            text/markdown: {}
            application/json: {}
            application/x-ndjson: {}
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/conversations/{id}/export:
    get:
      tags: [Import and export]
      summary: Download one conversation
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: format, in: query, schema: { type: string, enum: [md, json], default: md } }
        - { name: attachments, in: query, schema: { type: boolean, default: true }, description: "JSON only: inline attachments as base64" }
        - { name: tz, in: query, schema: { type: string } }
      responses:
        "200":
          description: The conversation
          content:
            text/markdown: {}
            application/json: {}
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/export:
    get:
      tags: [Import and export]
      summary: Download every conversation
      parameters:
        - { name: format, in: query, schema: { type: string, enum: [json, md], default: json }, description: "md gives a zip of Markdown files" }
        - { name: provider, in: query, schema: { $ref: "#/components/schemas/ChatProviderId" } }
        - { name: attachments, in: query, schema: { type: boolean, default: false } }
        - { name: tz, in: query, schema: { type: string } }
      responses:
        "200":
          description: The export
          content:
            application/json: {}
            application/zip: {}
  /v1/import:
    post:
      tags: [Import and export]
      summary: Import exports from ChatGPT, Claude, Gemini, AI Studio, Claude Code or Nightshift
      parameters:
        - { name: preview, in: query, schema: { type: boolean, default: false }, description: "Return what was found and an id to commit, instead of importing" }
        - { name: gap_minutes, in: query, schema: { type: integer, default: 30, minimum: 1, maximum: 1440 }, description: "Gemini Takeout: prompts closer than this form one conversation" }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - type: object
                  required: [files]
                  properties:
                    files: { type: array, items: { type: string, format: binary } }
                - $ref: "#/components/schemas/ImportOptions"
      responses:
        "201":
          description: With preview, an ImportPreview; otherwise an ImportResult
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ImportPreview"
                  - $ref: "#/components/schemas/ImportResult"
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/import/{id}/commit:
    post:
      tags: [Import and export]
      summary: Import a previewed upload
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ImportOptions" }
      responses:
        "200":
          description: What was created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportResult" }
        "404": { description: "The preview expired" }
  /v1/import/{id}:
    delete:
      tags: [Import and export]
      summary: Discard a preview
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: "Discarded" }
  /v1/uploads:
    post:
      tags: [Attachments]
      summary: Upload files
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "201":
          description: Stored attachments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Attachment" }
  /v1/uploads/base64:
    post:
      tags: [Attachments]
      summary: Upload one file as base64, easier from a script
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, data]
              properties:
                name: { type: string }
                data: { type: string, description: "Base64 file contents" }
                mime: { type: string }
      responses:
        "201":
          description: Stored attachment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attachment" }
  /v1/uploads/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [Attachments]
      summary: Download an attachment
      responses:
        "200": { description: "The file" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Attachments]
      summary: Delete an attachment no message refers to
      responses:
        "200": { description: "Deleted, or kept because it is still in use" }
  /v1/messages:
    post:
      tags: [Messages]
      summary: Queue one message, or a chain that runs in order
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MessageRequest" }
            examples:
              single:
                summary: One message, sent now
                value:
                  provider: openai
                  conversation_id: 8f2c…
                  text: Summarise today's incidents
              whenLimitResets:
                summary: A Claude Code chain that starts when the limit resets
                value:
                  provider: claude-code
                  cwd: /srv/projects/api
                  mode: reset
                  permission_mode: acceptEdits
                  webhook_url: https://example.com/hook
                  messages:
                    - text: Run the test suite and fix what fails.
                    - text: Then write the changelog entry.
      responses:
        "201":
          description: The jobs that were created, in order
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items: { $ref: "#/components/schemas/Job" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/jobs:
    get:
      tags: [Messages]
      summary: List queued and finished messages
      parameters:
        - { name: status, in: query, schema: { type: string }, description: "Comma separated, or \"active\"" }
        - { name: provider, in: query, schema: { $ref: "#/components/schemas/ProviderId" } }
        - { name: conversation_id, in: query, schema: { type: string } }
        - { name: session_id, in: query, schema: { type: string } }
        - { name: thread_key, in: query, schema: { type: string } }
        - { name: since, in: query, schema: { type: integer }, description: "Only jobs updated at or after this epoch ms" }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
      responses:
        "200":
          description: Oldest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Job" }
  /v1/jobs/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [Messages]
      summary: One message, optionally waiting for it to finish
      parameters:
        - name: wait
          in: query
          schema: { type: integer, maximum: 300 }
          description: Seconds to hold the request open until the message reaches a final state
      responses:
        "200":
          description: The message
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Messages]
      summary: Edit a message that has not been sent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text: { type: string }
                files: { type: array, items: { type: string } }
                model: { type: string }
                effort: { $ref: "#/components/schemas/Effort" }
                max_tokens: { type: integer }
                permission_mode: { $ref: "#/components/schemas/PermissionMode" }
                mode: { $ref: "#/components/schemas/Mode" }
                run_at: { type: [integer, string] }
                delay_seconds: { type: integer }
                retry_on_limit: { type: boolean }
                webhook_url: { type: string, format: uri }
                repeat: { type: string, enum: [none, daily, weekdays, weekly], description: "First message of a chain only. none stops repeating" }
                timezone: { type: string }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "400": { $ref: "#/components/responses/BadRequest" }
    delete:
      tags: [Messages]
      summary: Delete a message
      parameters:
        - { name: cascade, in: query, schema: { type: boolean, default: true }, description: "Re-link anything chained behind it" }
        - { name: force, in: query, schema: { type: boolean, default: false }, description: "Cancel it first if it is sending" }
      responses:
        "200": { description: "Deleted" }
  /v1/jobs/{id}/cancel:
    post:
      tags: [Messages]
      summary: Stop a message, including one that is sending
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The cancelled message
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
  /v1/jobs/{id}/retry:
    post:
      tags: [Messages]
      summary: Send now, ignoring a known limit
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The re-queued message
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: An API token created in Settings. `X-API-Key` is also accepted.
  responses:
...
    NotFound:
      description: No such thing
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    ProviderId:
      type: string
      enum: [anthropic, claude-code, openai, gemini, ollama]
    ChatProviderId:
      type: string
      enum: [anthropic, openai, gemini, ollama]
    Effort:
      type: string
      enum: ["", minimal, low, medium, high, xhigh, max]
...
    Conversation:
      type: object
      properties:
        id: { type: string }
        provider: { $ref: "#/components/schemas/ChatProviderId" }
        title: { type: string }
        model: { type: string }
        system: { type: string }
        messageCount: { type: integer }
        createdAt: { type: integer }
        updatedAt: { type: integer }
        source:
          type: object
          description: Set on imported conversations
          properties:
            kind: { type: string, enum: [chatgpt, claude, gemini, aistudio, claude-code, nightshift, backup] }
            id: { type: string }
...